# app/Http/Controllers/Api/PreInscricaoController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Models\PreInscricao;
use Illuminate\Http\Request;

/**
 * @OA\Tag(
 *     name="Pré-inscrições",
 *     description="Operações relacionadas às pré-inscrições de cursos"
 * )
 */
class PreInscricaoController extends Controller
{
   
    /**
     * @OA\Post(
     *     path="/pre-inscricoes",
     *     tags={"Pré-inscrições"},
     *     summary="Realizar uma nova pré-inscrição",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(ref="#/components/schemas/PreInscricaoInput")
     *     ),
     *     @OA\Response(
     *         response=201,
     *         description="Pré-inscrição realizada com sucesso",
     *         @OA\JsonContent(ref="#/components/schemas/PreInscricao")
     *     )
     * )
     */
    public function store(Request $request)
    {
        $validated = $request->validate([
            'curso_id' => 'required|exists:cursos,id',
            'centro_id' => 'required|exists:centros,id',
            'horario_id' => 'nullable|exists:horarios,id',
            'nome_completo' => 'required|string|max:100',
            'contactos' => [
                'required',
                'array',
                'min:1'
            ],
            'contactos.*' => [
                'required',
                'string',
                'regex:/^9\d{8}$/'
            ],
            'email' => 'nullable|email|max:100',
            'observacoes' => 'nullable|string|max:500'
        ]);

        // Formatar dados
        $validated['contactos'] = array_map('strval', $validated['contactos']);
        $validated['status'] = 'pendente'; // Status padrão
        
        // Normalizar email para lowercase se fornecido
        if (!empty($validated['email'])) {
            $validated['email'] = strtolower($validated['email']);
        }

        $preInscricao = PreInscricao::create($validated);

        return response()->json([
            'status' => 'sucesso',
            'mensagem' => 'Pré-inscrição realizada!',
            'dados' => $preInscricao
        ], 201);
    }



    /**
     * @OA\Get(
     *     path="/pre-inscricoes",
     *     tags={"Pré-inscrições"},
     *     summary="Listar todas as pré-inscrições",
     *     @OA\Response(
     *         response=200,
     *         description="Lista de pré-inscrições",
     *         @OA\JsonContent(type="array", @OA\Items(ref="#/components/schemas/PreInscricao"))
     *     )
     * )
     */
    public function index()
    {
        $preInscricoes = PreInscricao::with(['curso', 'centro', 'horario'])->get();
        return response()->json($preInscricoes);
    }


    /**
     * @OA\Put(
     *     path="/pre-inscricoes/{id}",
     *     tags={"Pré-inscrições"},
     *     summary="Atualizar status da pré-inscrição (apenas status)",
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         required=true,
     *         @OA\Schema(type="integer")
     *     ),
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(ref="#/components/schemas/PreInscricaoStatusInput")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Status atualizado",
     *         @OA\JsonContent(ref="#/components/schemas/PreInscricao")
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Pré-inscrição não encontrada"
     *     )
     * )
     */
    public function update(Request $request, $id)
    {
        $preInscricao = PreInscricao::find($id);

        if (!$preInscricao) {
            return response()->json([
                'status' => 'erro',
                'mensagem' => 'Pré-inscrição não encontrada!'
            ], 404);
        }

        // Só permite editar o status
        $validated = $request->validate([
            'status' => 'required|in:pendente,confirmado,cancelado'
        ]);
        $preInscricao->update($validated);

        return response()->json([
            'status' => 'sucesso',
            'mensagem' => 'Status atualizado!',
            'dados' => $preInscricao
        ]);
    }


    /**
     * @OA\Delete(
     *     path="/pre-inscricoes/{id}",
     *     tags={"Pré-inscrições"},
     *     summary="Deletar pré-inscrição",
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         required=true,
     *         @OA\Schema(type="integer")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Pré-inscrição deletada"
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Pré-inscrição não encontrada"
     *     )
     * )
     */
    public function destroy($id)
    {
        $preInscricao = PreInscricao::find($id);
        if (!$preInscricao) {
            return response()->json([
                'status' => 'erro',
                'mensagem' => 'Pré-inscrição não encontrada!'
            ], 404);
        }
        $preInscricao->delete();
        return response()->json([
            'status' => 'sucesso',
            'mensagem' => 'Pré-inscrição deletada com sucesso!'
        ]);
    }


    /**
     * @OA\Get(
     *     path="/pre-inscricoes/{id}",
     *     tags={"Pré-inscrições"},
     *     summary="Buscar pré-inscrição por ID",
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         required=true,
     *         @OA\Schema(type="integer")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Pré-inscrição encontrada",
     *         @OA\JsonContent(ref="#/components/schemas/PreInscricao")
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Pré-inscrição não encontrada"
     *     )
     * )
     */
    public function show($id)
    {
        $preInscricao = PreInscricao::with(['curso', 'centro', 'horario'])->find($id);
        if (!$preInscricao) {
            return response()->json([
                'status' => 'erro',
                'mensagem' => 'Pré-inscrição não encontrada!'
            ], 404);
        }
        return response()->json([
            'status' => 'sucesso',
            'dados' => $preInscricao
        ]);
    }
}